third_party/blink/web_tests/http/tests/websocket/multiple-connections-throttled-expected.txt

Test that WebSocket connections are throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL openedCount should be 255. Was 300.
PASS successfullyParsed is true

TEST COMPLETE
